algo work and shift be gone
